Create Answer Selection Function
handleSelected receives the question and answer id from Quiz. It
then iterates over the array of questions and where there is a match
for the question id switches the state of isSelected for the clicked
answer and sets the isSelected state of all other answers for that
question to false.