Add checkScore function
checkScore passed to onClick of button. Function iterates over
questions to check for the selected answer matching the correct answer
and if so increments the score state. The function also flips the
scoreQuiz state to true to render the final score display and the
dynamic styling of the answer buttons.

The "correct"field has been removed from "questions" as it proved
unecessary.